added front and backend ChatGPT files and did some cleanup
