--- client/src/PLAYindex.js
@@ -1,33 +1,0 @@
-import axios from "axios";
-
-function App() {
-  const [prompt, setPrompt] = useState("");
-  const [response, setResponse] = useState("");
-
-  const handleSubmit = (e) => {
-    e.preventDefault();
-
-    axios
-      .post("/chat", { prompt })
-      .then((res) => {
-        setResponse(res.data);
-      })
-      .catch((err) => {
-        console.error(err);
-      });
-  };
-
-  return (
-    <div>
-      <form onSubmit={handleSubmit}>
-        <input
-          type="text"
-          value={prompt}
-          onChange={(e) => setPrompt(e.target.value)}
-        />
-        <button type="submit">Submit</button>
-      </form>
-      <p>{response}</p>
-    </div>
-  );
-}
