--- client/src/components/AskChatGPT/index_main.js
@@ -1,15 +1,0 @@
-import React from 'react';
-//import { useLocation, useNavigate } from 'react-router-dom';
-
-const AskChatGPT = () => {
-  //const location = useLocation();
-  //const navigate = useNavigate();
-  return (
-    <div>
-    <h1>Need help coming up with steps? Ask ChatGPT!</h1>
-    <div id="chat-container"></div>
-    </div>
-  );
-};
-
-export default AskChatGPT;
